Implement file submission upon translate button click for both url inputs and image upload inputs
