+++ src/styles/index.css
@@ -22,6 +22,15 @@ body,
   justify-content: center;
 }
 
+.input-page {
+  height: 100%;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
 .input-container {
   display: flex;
   flex-direction: column-reverse;
@@ -105,12 +114,11 @@ div.top:hover {
   height: 20px;
   font-size: .8em;
   box-shadow: 0px 0px 3px rgba(0, 0, 0, 0.2);
-  transition: box-shadow .5s ease-in;
+  transition: box-shadow .2s ease-in;
 }
 
 .input:focus {
-  box-shadow: 0px 0px 6px rgba(0, 0, 0, 0.2);
-  /*transition: box-shadow .5s ease-in;*/
+  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.2);
 }
 
 .input::-webkit-input-placeholder {
